requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.11.0
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.7
easyocr==1.7.1
environ==1.0
exceptiongroup==1.2.1
filelock==3.15.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.0
golang==0.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
inflection==0.5.1
intel-openmp==2021.4.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.3.2
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mysqlclient==2.2.4
networkx==3.3
ngrok==1.3.0
ninja==1.11.1.1
numpy==2.0.0
openai==0.28.0
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
pyparsing==3.0.3
pyperclip==1.9.0
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
responses==0.25.3
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.13.1
selenium==4.21.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.12.0
tifffile==2024.6.18
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.2
webdriver-manager==4.0.1
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.9.4
